extensionfacade: Remove use of finalize
Using finalize is not recommended and is being planned for removal.
We update the service tracking code to close the ServiceTracker when
it detects the parent object has been garbage collected when called
at the customizer methods.
